# Conflicts:
#	app/Http/Controllers/Api/TourComponentController.php
#	app/Listeners/CheckoutSuccessfulListener.php
#	app/Models/AccommodationGroup.php
#	app/Models/Order.php
#	app/Models/Order/Payment/PaymentIntention.php
#	app/Models/PaymentReminder.php
#	app/Repository/BespokeReportRepository.php
#	app/Repository/OrderRepository.php
#	app/Repository/ReportRepository.php
#	resources/views/pages/tour/view.blade.php
#	routes/web.php

+++ app/Http/Controllers/Models/FlightInventoryTourController.php
@@ -55,9 +55,26 @@ class FlightInventoryTourController extends Controller
         return redirect()->route('tours.view', ['tour' => $tour,]);
     }
 
+    public function restore(Tour $tour, $flightInventoryTour)
+    {
+        $inventoryTour = FlightInventoryTour::withTrashed()->find($flightInventoryTour);
+        if ($inventoryTour->trashed()) {
+            $inventoryTour->restore();
+        } else {
+            $inventoryTour->is_bookable = true;
+            $inventoryTour->save();
+        }
+        return redirect()->route('tours.view', ['tour' => $tour,]);
+    }
+
     public function destroy(Tour $tour, FlightInventoryTour $flightInventoryTour)
     {
-        $flightInventoryTour->delete();
+        if ($tour->orders()->count() > 0) {
+            $flightInventoryTour->is_bookable = false;
+            $flightInventoryTour->save();
+        } else {
+            $flightInventoryTour->delete();
+        }
         return redirect()->route('tours.view', ['tour' => $tour,]);
     }
 }
